functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure.storage.blob
anyio==4.1.0
azure-common==1.1.28
azure-core==1.29.6
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
exceptiongroup==1.2.0
Flask==3.0.0
gunicorn==21.2.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
openai==1.6.1


# azure-functions
# azure.storage.blob
# azure-search-documents
# openai
# azure-search-documents
# azure-core
# requests
# asyncio




